# openapi.yaml
openapi: 3.0.3
info:
  title: Tucano API
  version: 0.3.0
  description: >-
    API do marketplace de profissionais autônomos. A plataforma facilita descoberta,
    compatibilidade e contato direto; não processa o pagamento dos serviços.
servers:
  - url: /api
paths:
...
  /professional/dashboard:
    get:
      tags: [professional]
      summary: Endpoint protegido por perfil profissional
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Acesso profissional autorizado}
        "403": {description: Perfil profissional necessário}
...
  /professional/profile:
    get:
      tags: [professional]
      summary: Retorna o perfil profissional privado da conta
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Perfil atual, content: {application/json: {schema: {type: object, properties: {profile: {$ref: "#/components/schemas/ProfessionalProfile"}}}}}}
        "403": {description: Perfil profissional necessário}
    put:
      tags: [professional]
      summary: Atualiza apresentação, categoria, região e disponibilidade
      description: Coordenadas são armazenadas para cálculo geoespacial e nunca retornadas.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/ProfessionalProfileInput"}}}
      responses:
        "200": {description: Perfil atualizado}
        "422": {$ref: "#/components/responses/ValidationError"}
  /professional/portfolio:
    post:
      tags: [professional]
      summary: Adiciona uma imagem ao portfólio básico
      description: Aceita JPEG ou PNG com até 5 MB; o limite técnico atual é de 12 imagens.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties: {file: {type: string, format: binary}}
      responses:
        "201": {description: Imagem adicionada}
        "413": {description: Arquivo grande demais}
        "415": {description: Tipo de mídia não permitido}
  /professional/portfolio/{itemId}:
    delete:
      tags: [professional]
      summary: Remove uma imagem pertencente ao profissional
      security: [{bearerAuth: []}]
      parameters:
        - {in: path, name: itemId, required: true, schema: {type: string, format: uuid}}
      responses:
        "204": {description: Imagem removida}
        "404": {description: Item não encontrado ou não pertencente à conta}
  /professionals/search:
    get:
      tags: [discovery]
      summary: Busca profissionais relevantes por categoria e distância
      description: Retorna somente região declarada e distância arredondada, sem coordenadas.
      security: [{bearerAuth: []}]
      parameters:
        - {in: query, name: category_id, required: true, schema: {type: string, format: uuid}}
        - {in: query, name: latitude, required: true, schema: {type: number, minimum: -90, maximum: 90}}
        - {in: query, name: longitude, required: true, schema: {type: number, minimum: -180, maximum: 180}}
        - {in: query, name: radius_km, schema: {type: integer, minimum: 1, maximum: 200, default: 25}}
        - {in: query, name: available_now, schema: {type: boolean, default: false}}
        - {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
        - {in: query, name: page_size, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
      responses:
        "200": {description: Página de profissionais, content: {application/json: {schema: {$ref: "#/components/schemas/ProfessionalSearchPage"}}}}
        "403": {description: Perfil contratante necessário}
        "422": {$ref: "#/components/responses/ValidationError"}
  /uploads/{storageKey}:
    get:
      tags: [discovery]
      summary: Serve uma imagem pública do portfólio por chave não previsível
      parameters:
        - {in: path, name: storageKey, required: true, schema: {type: string}}
      responses:
        "200": {description: Imagem JPEG ou PNG}
        "404": {description: Arquivo não encontrado}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationError:
      description: Dados inválidos
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
...
  schemas:
...
    ServiceCategory:
      type: object
      required: [id, slug, name, description]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        name: {type: string}
        description: {type: string}
    AvailabilityInput:
      type: object
      required: [weekday, start_time, end_time]
      properties:
        weekday: {type: integer, minimum: 0, maximum: 6, description: Domingo é 0 e sábado é 6.}
        start_time: {type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$"}
        end_time: {type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$"}
    ProfessionalProfileInput:
      type: object
      required: [bio, primary_category_id, skills, service_region, service_radius_km, available_now, availability_timezone, availabilities]
      properties:
        bio: {type: string, maxLength: 600}
        primary_category_id: {type: string, format: uuid}
        skills: {type: array, maxItems: 20, items: {type: string, maxLength: 50}}
        service_region: {type: string, minLength: 2, maxLength: 120}
        latitude: {type: number, minimum: -90, maximum: 90, writeOnly: true}
        longitude: {type: number, minimum: -180, maximum: 180, writeOnly: true}
        service_radius_km: {type: integer, minimum: 1, maximum: 200}
        reference_price_cents: {type: integer, nullable: true, minimum: 0}
        available_now: {type: boolean}
        availability_timezone: {type: string, example: America/Sao_Paulo}
        availabilities: {type: array, maxItems: 21, items: {$ref: "#/components/schemas/AvailabilityInput"}}
    ProfessionalProfile:
      type: object
      required: [id, bio, skills, service_region, location_configured, service_radius_km, available_now, availability_timezone, availabilities, portfolio]
      properties:
        id: {type: string, format: uuid}
        bio: {type: string}
        primary_category: {allOf: [{$ref: "#/components/schemas/ServiceCategory"}], nullable: true}
        skills: {type: array, items: {type: string}}
        service_region: {type: string}
        location_configured: {type: boolean}
        service_radius_km: {type: integer}
        reference_price_cents: {type: integer, nullable: true}
        available_now: {type: boolean}
        availability_timezone: {type: string}
        availabilities: {type: array, items: {$ref: "#/components/schemas/AvailabilityInput"}}
        portfolio: {type: array, items: {$ref: "#/components/schemas/PortfolioItem"}}
    PortfolioItem:
      type: object
      required: [id, url, original_name, media_type, size_bytes, sort_order]
      properties:
        id: {type: string, format: uuid}
        url: {type: string}
        original_name: {type: string}
        media_type: {type: string, enum: [image/jpeg, image/png]}
        size_bytes: {type: integer}
        sort_order: {type: integer}
    ProfessionalSearchResult:
      type: object
      description: Não contém latitude, longitude nem a coluna geoespacial armazenada.
      required: [id, display_name, category, skills, service_region, service_radius_km, available_now, distance_km]
      properties:
        id: {type: string, format: uuid}
        display_name: {type: string}
        bio: {type: string}
        category: {$ref: "#/components/schemas/ServiceCategory"}
        skills: {type: array, items: {type: string}}
        service_region: {type: string}
        service_radius_km: {type: integer}
        reference_price_cents: {type: integer, nullable: true}
        available_now: {type: boolean}
        distance_km: {type: number, format: double}
        cover_url: {type: string}
    ProfessionalSearchPage:
      type: object
      required: [items, page, page_size, total, total_pages]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/ProfessionalSearchResult"}}
        page: {type: integer}
        page_size: {type: integer}
        total: {type: integer}
        total_pages: {type: integer}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}